+++ qiskit/aqua/components/oracles/__init__.py
@@ -18,10 +18,12 @@
 from .oracle import Oracle
 from .truth_table_oracle import TruthTableOracle
 from .logical_expression_oracle import LogicalExpressionOracle
+from .custom_circuit_oracle import CustomCircuitOracle
 
 
 __all__ = [
     'Oracle',
     'TruthTableOracle',
     'LogicalExpressionOracle',
+    'CustomCircuitOracle',
 ]