+++ qiskit/aqua/components/uncertainty_models/multivariate_distribution.py
@@ -36,15 +36,15 @@ class MultivariateDistribution(UncertaintyModel, ABC):
     def get_section_key_name(cls):
         return Pluggable.SECTION_KEY_MULTIVARIATE_DISTRIBUTION
 
-    def __init__(self, num_qubits, probabilities, low, high):
+    def __init__(self, num_qubits, low, high, probabilities=None):
         """
-        Constructor
+        Constructor.
 
         Args:
-            num_qubits (:obj:`list` of :obj:`list`): assigns qubits to dimensions
-            probabilities: map - maps index tuples to probabilities
-            low (list): lowest value per dimension
-            high (list): highest value per dimension
+            num_qubits (array or list): assigns qubits to dimensions
+            probabilities (map): map - maps index tuples to probabilities
+            low (array or list): lowest value per dimension
+            high (array or list): highest value per dimension
         """
 
         # derive dimension from qubit assignment
@@ -59,22 +59,34 @@ class MultivariateDistribution(UncertaintyModel, ABC):
         # call super constructor
         super().__init__(num_target_qubits)
 
-        # normalize probabilities
-        probabilities = np.asarray(probabilities)
-        probabilities = probabilities / np.sum(probabilities)
-
         self._num_values = []
         for i in range(self._dimension):
             self._num_values += [2 ** num_qubits[i]]
-        self._probabilities = probabilities
-        self._probabilities_vector = np.reshape(probabilities, 2**num_target_qubits)
 
-        self._probabilities_vector = np.asarray(self._probabilities_vector)
-        self._low = low
-        self._high = high
-        self._values = []
-        for i in range(self._dimension):
-            self._values += [np.linspace(self._low[i], self._high[i], self._num_values[i])]
+        if probabilities is not None:
+
+            # normalize probabilities
+            probabilities = np.asarray(probabilities)
+            probabilities = probabilities / np.sum(probabilities)
+
+            self._probabilities = probabilities
+            self._probabilities_vector = np.reshape(probabilities, 2**num_target_qubits)
+            self._probabilities_vector = np.asarray(self._probabilities_vector)
+        else:
+            self._probabilities = None
+            self._probabilities_vector = None
+
+        if low is not None:
+            self._low = low
+        else:
+            self._low = np.zeros(self.dimension)
+
+        if high is not None:
+            self._high = high
+        else:
+            self._high = np.zeros(self.dimension)
+            for i in range(self.dimension):
+                self._high[i] = 2**num_qubits[i] - 1
 
     @property
     def num_qubits(self):
@@ -107,12 +119,6 @@ class MultivariateDistribution(UncertaintyModel, ABC):
     @property
     def probabilities_vector(self):
         return self._probabilities_vector
-
-    def required_ancillas(self):
-        return 0
-
-    def required_ancillas_controlled(self):
-        return 0
 
     def build(self, qc, q, q_ancillas=None, params=None):
         custom_state = Custom(self.num_target_qubits, state_vector=np.sqrt(self._probabilities_vector))