+++ CHANGELOG.rst
@@ -32,19 +32,24 @@ Added
 - Support to use arbitrary mixer Hamiltonian in ``QAOA``. This allows to use QAOA in constrained optimization problems [arXiv:1709.03489].
 - Added variational algorithm base class ``VQAlgorithm``, implemented by ``VQE`` and ``QSVMVariational``.
 - Added ``ising/docplex.py`` for automatically generating Ising Hamiltonian from optimization models of DOcplex.
+- Added ``'basic-dirty-ancilla'`` mode for ``mct``.
 - Added ``mcmt`` for Multi-Controlled, Multi-Target gate.
 - Exposed capabilities to generate circuits from logical AND, OR, DNF (disjunctive normal forms), and CNF (conjunctive normal forms) formulae.
 - Added the capability to generate circuits from ESOP (exclusive sum of products) formulae with optional optimization based on Quine-McCluskey and ExactCover.
 - Added ``LogicalExpressionOracle`` for generating oracle circuits from arbitrary boolean logic expressions (including DIMACS support) with optional optimization capability.
 - Added ``TruthTableOracle`` for generating oracle circuits from truth-tables with optional optimization capability.
+- Added ``CustomCircuitOracle`` for generating oracle from user specified circuits.
 - Added implementation of the Deutsch-Josza algorithm.
 - Added implementation of the Bernstein-Vazirani algorithm.
 - Added implementation of the Simon's algorithm.
+- Added implementation of the Shor's algorithm.
 - Added optional capability for ``Grover``'s algorithm to take a custom initial state (as opposed to the default uniform superposition)
 - Added capability to create a ``Custom`` initial state using existing circuit.
 - Added the ADAM (and AMSGRAD) optimization algorithm
 - Multivariate distributions added, so uncertainty models now have univariate and multivariate distribution components
 - Added option to include or skip the swaps operations for qft and iqft circuit constructions.
+- Added classical linear system solver ``ExactLSsolver``
+- Added parameters ``auto_hermitian`` and ``auto_resize`` to ``HHL`` algorithm to support non-hermititan and non 2**n sized matrices by default
 
 Removed
 -------